Normalize herder inclusion_fee fields from u64 to i64
Change QueuedTransaction.inclusion_fee, FeeEntry.inclusion_fee,
QueueEntry.inclusion_fee, and EvictionThresholds fee tuples from u64
to i64 to align with stellar-core's int64_t representation and
eliminate ~15 redundant `as i64` casts at fee_rate_cmp call sites.

Also update compute_better_fee, min_inclusion_fee_to_beat, and
can_replace_by_fee signatures to use i64, and add boundary tests
for i64::MAX saturation and large-value cross-multiplication.

Closes #1832

+++ crates/herder/src/surge_pricing.rs
@@ -245,7 +245,7 @@ impl EvictionExclusion {
 /// with tie-breaking based on a deterministic hash to ensure consistent ordering.
 #[derive(Clone)]
 pub(crate) struct QueueEntry {
-    inclusion_fee: u64,
+    inclusion_fee: i64,
     op_count: u32,
     tie_breaker: [u8; 32],
     hash: [u8; 32],
@@ -287,11 +287,10 @@ impl PartialOrd for QueueEntry {
 
 impl Ord for QueueEntry {
     fn cmp(&self, other: &Self) -> Ordering {
-        // inclusion_fee is u64 validated non-negative at construction (from XDR i64).
         let ord = fee_rate_cmp(
-            self.inclusion_fee as i64,
+            self.inclusion_fee,
             self.op_count,
-            other.inclusion_fee as i64,
+            other.inclusion_fee,
             other.op_count,
         );
         if ord != Ordering::Equal {
@@ -713,9 +712,9 @@ impl SurgePricingPriorityQueue {
             };
 
             if fee_rate_cmp(
-                entry.inclusion_fee as i64,
+                entry.inclusion_fee,
                 entry.op_count,
-                tx.inclusion_fee as i64,
+                tx.inclusion_fee,
                 tx.op_count,
             ) != Ordering::Less
             {
